Fix a few bugs. Seems mostly usable now.

+++ storage-service/src/main/kotlin/dk/sdu/cloud/file/services/linuxfs/CLibrary.kt
@@ -12,8 +12,8 @@ interface CLibrary : Library {
     fun setxattr(path: String, name: String, value: ByteArray, size: Int, position: Int): Int
     fun listxattr(path: String, destination: ByteArray, length: Int): Int
     fun removexattr(path: String, name: String): Int
-    fun setfsuid(uid: Int): Int
-    fun setfsgid(uid: Int): Int
+    fun setfsuid(uid: Long): Int
+    fun setfsgid(uid: Long): Int
 
     companion object {
         val INSTANCE =
